requirements.txt
absl-py==0.12.0
addict==2.4.0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
blinker==1.4
bravado==11.0.3
bravado-core==5.17.0
breezy==3.0.2
brz-debian==3.0.0
cachetools==4.2.1
carla==0.9.11
catfish==1.4.13
certifi==2019.11.28
cffi==1.14.5
Chameleon==3.6.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
croniter==0.3.29
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.7
devscripts===2.20.2ubuntu2
dill==0.3.3
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dulwich==0.19.15
entrypoints==0.3
fastimport==0.9.8
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gpg===1.13.1-unknown
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
imageio==2.9.0
importlib-metadata==1.5.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.10.1
joblib==1.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
lptools==0.2.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.4.1
meld==3.20.2
menulibre==2.2.1
mistune==0.8.4
monotonic==1.6
more-itertools==4.2.0
msgpack==0.6.2
mugshot==0.4.2
multiprocess==0.70.11.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
neptune-client==0.9.4
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
open3d==0.12.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pathos==0.2.7
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plyfile==0.7.3
pox==0.2.9
ppft==1.6.6.3
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.5.1
ptyprocess==0.7.0
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycryptodomex==3.6.1
pycups==1.9.73
pycurl==7.43.0.2
pygame==2.0.1
PyGithub==1.43.7
Pygments==2.8.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySimpleSOAP==1.16.2
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-debianbts==3.0.2
python-gitlab==2.0.1
python-gnupg==0.4.5
python-magic==0.4.16
python-networkmanager==2.1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportbug===7.6.0ubuntu1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3987==1.3.8
rsa==4.7.2
salt==3003
scikit-image==0.18.1
scikit-learn==0.22.2.post1
scipy==1.6.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sgt-launcher==0.2.5
Shapely==1.7.1
simplejson==3.16.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
ssh-import-id==5.10
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-dev-tools==0.176
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.5
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xcffib==0.8.1
xkit==0.0.0
zipp==1.0.0
